extract header component
